refactor: route take_message through Dispatcher; delete legacy messages.py
receptionist/messages.py content is now covered by:
  - receptionist/messaging/models.py (Message dataclass)
  - receptionist/messaging/channels/file.py (file-write logic)
  - receptionist/messaging/dispatcher.py (multi-channel fan-out)

Receptionist.take_message now constructs a Dispatcher from the
per-business channel list and dispatches. A synchronous failure
returns a caller-visible fallback to the LLM.
