+++ tests/test_messages.py
@@ -1,51 +1,24 @@
 # tests/test_messages.py
-import json
-from pathlib import Path
-from receptionist.messages import save_message, Message
+"""Legacy smoke tests — the full coverage lives in tests/messaging/test_file_channel.py."""
+from __future__ import annotations
+
+import pytest
+
+from receptionist.messaging.models import Message
 
 
-def test_save_message_creates_file(tmp_path):
-    msg = Message(
-        caller_name="John Doe",
-        callback_number="+15559876543",
-        message="Please call me back about my appointment.",
-        business_name="Test Dental",
-    )
-    save_message(msg, delivery="file", file_path=str(tmp_path))
-
-    files = list(tmp_path.glob("*.json"))
-    assert len(files) == 1
-
-    data = json.loads(files[0].read_text())
-    assert data["caller_name"] == "John Doe"
-    assert data["callback_number"] == "+15559876543"
-    assert data["message"] == "Please call me back about my appointment."
-    assert data["business_name"] == "Test Dental"
-    assert "timestamp" in data
+def test_message_timestamp_autofills():
+    msg = Message("Jane", "+15551112222", "Call me", "Acme")
+    assert msg.timestamp  # auto-populated ISO timestamp
 
 
-def test_save_multiple_messages(tmp_path):
-    for i in range(3):
-        msg = Message(
-            caller_name=f"Caller {i}",
-            callback_number=f"+1555000000{i}",
-            message=f"Message {i}",
-            business_name="Test Dental",
-        )
-        save_message(msg, delivery="file", file_path=str(tmp_path))
-
-    files = list(tmp_path.glob("*.json"))
-    assert len(files) == 3
-
-
-def test_save_message_creates_directory(tmp_path):
-    nested = tmp_path / "sub" / "dir"
-    msg = Message(
-        caller_name="Jane",
-        callback_number="+15551111111",
-        message="Test",
-        business_name="Test Dental",
-    )
-    save_message(msg, delivery="file", file_path=str(nested))
-    assert nested.exists()
-    assert len(list(nested.glob("*.json"))) == 1
+def test_message_to_dict_roundtrip():
+    msg = Message("Jane", "+15551112222", "Call me", "Acme", timestamp="2026-01-01T00:00:00+00:00")
+    d = msg.to_dict()
+    assert d == {
+        "caller_name": "Jane",
+        "callback_number": "+15551112222",
+        "message": "Call me",
+        "business_name": "Acme",
+        "timestamp": "2026-01-01T00:00:00+00:00",
+    }
